Moves .arc copy into module, improves progress ticks in deploy
