+++ src/hydrate/shared/_copy.js
@@ -1,75 +1,61 @@
+let assert = require('@smallwins/validate/assert')
+let copyArc = require('./_arc')
 let cp = require('cpr')
 let exists = require('path-exists').sync
-let fs = require('fs')
 let lambdaPath = require('../../util/get-lambda-name')
 let parallel = require('run-parallel')
-let parse = require('@architect/parser')
 let series = require('run-series')
 
 module.exports = function copyCommon(params, callback) {
-  let {arc, pathToCode, tick} = params
 
+  assert(params, {
+    arc: Object,
+    pathToCode: Array,
+    // tick: Function,
+  })
+
+  let { arc, pathToCode, tick } = params
+  
   series([
     function _shared(callback) {
       if (tick) tick(`Copying src/shared into Functions...`)
-
       let src = process.cwd() + '/src/shared'
 
-      // Bail early if project doesn't use src/shared
-      if (!exists(src)) {
-        callback()
-      }
-      else {
-        parallel(pathToCode.map(path => {
-          return function _copy(callback) {
-            let dest = process.cwd() + '/' + path + '/node_modules/@architect/shared'
-            let arcFileDest = dest + '/.arc'
+      // Skip copying src/shared early if project doesn't use it
+      let shared = false
+      if (exists(src)) shared = true
+      parallel(pathToCode.map(path => {
+        return function _copy(callback) {
+          let dest = process.cwd() + '/' + path + '/node_modules/@architect/shared'
 
-            series([
-              // Copy src/shared to all functions
-              function copyShared(callback) {
+          series([
+            // Maybe copy src/shared to all Functions
+            function copyShared(callback) {
+              if (shared) {
                 copy(src, dest, callback)
-              },
-              function copyArc(callback) {
-                // Possible .arc manifest locations
-                let arcFileSrc = process.cwd() + '/.arc'
-                let arcAppDotArcPath = process.cwd() + '/app.arc'
-                let arcYamlPath = process.cwd() + '/arc.yaml'
-                let arcJsonPath = process.cwd() + '/arc.json'
-                if (exists(arcFileSrc)) {
-                  copy(arcFileSrc, arcFileDest, callback)
-                }
-                else if (exists(arcAppDotArcPath)) {
-                  copy(arcAppDotArcPath, arcFileDest, callback)
-                }
-                else if (exists(arcYamlPath)) {
-                  let raw = fs.readFileSync(arcYamlPath).toString()
-                  let arc = parse.yaml.stringify(raw)
-                  fs.writeFileSync(arcFileDest, arc)
-                  callback()
-                }
-                else if (exists(arcJsonPath)) {
-                  let raw = fs.readFileSync(arcJsonPath).toString()
-                  let arc = parse.json.stringify(raw)
-                  fs.writeFileSync(arcFileDest, arc)
-                  callback()
-                }
               }
-            ],
-            function done(err) {
-              if (err) callback(err)
-              else callback()
-            })
-          }
-        }),
-        function done(err) {
-          if (err) callback(err)
-          else {
-            if (tick) tick('')
-            callback()
-          }
-        })
-      }
+              else {
+                callback()
+              }
+            },
+            // Copy .arc to all Functions
+            function copyArcFile(callback) {
+              copyArc(path, callback)
+            }
+          ],
+          function done(err) {
+            if (err) callback(err)
+            else callback()
+          })
+        }
+      }),
+      function done(err) {
+        if (err) callback(err)
+        else {
+          if (tick) tick('')
+          callback()
+        }
+      })
     },
     function _views(callback) {
       if (tick) tick(`Copying src/views into Functions...`)
@@ -78,6 +64,7 @@ module.exports = function copyCommon(params, callback) {
 
       // Bail early if project doesn't use src/views
       if (!exists(src)) {
+        if (tick) tick('')
         callback()
       }
       else {
